extracted map bottom sheet added debounce
