+++ translations/en.json
@@ -112,7 +112,7 @@
     "BottomSheet": {
       "title": "Where are you parking?",
       "showDetails": "Show details",
-      "noZoneSelected": "No parking zones found. Use search or place the pin to parking zone nearby."
+      "noZoneSelected": "No parking zones found. Use search or place the pin to a parking zone nearby."
     }
   }
 }
